Changes to structure and shop
You can now change between battle and shop mode. Weapons will stop shooting in shop mode, and you can change the positions of the weapons there as well. Weapons states are kept between the different scene loads, they are spawned in by the WeaponManager at the start of the scene along side with the player ship.