Changes to various managers & Added TMP
Weapon movement is now done in the weapon class itself, each weapon is responsible to move to its desired position when it is changed through the weaponManager ( This took me to long to implement, I had a lot of bugs lol). Also added a small UI which switches between shop and battle scene. Weapons now also get spawned in the right slots when switching between the two scenes