Added small shop UI. Weapons can be dragged on ship
The shop now shows a random set of weapons which then can be dragged and put on a desired side of the ship. NOTE: There is some kind of bug that is still there where the data does not get read in a correct way. Will be fixed in the next commit. It took me way to long to figure out how to correctly save and read data. I did not understand the difference between an instantiated gameobject and the actual data of the weapon. Now the data is kept in the WeaponScriptableObject array and the gameobjects that are instantiated within each scene in the GameObject array weaponClones