Enemy now uses Target Joint 2D
Didn't realize I could use the joints to simulate some kind of gravity that would pull the enemies towards the center. The enemy script is now almost completely empty and the only that that is moving it is the target join component. For knockback, mass is now the thing that matters when a bullet collides with an enemy. My idea is to let the enemy get knocked back a tiny bit to make it look nice and also to make the game slightly easier.