Added Wave system
Waves store information about how many enemies will get spawned, which types of enemies get spawned, how they will get spawned, and how long the wave will be. Enemy class also had to be extended a bit to include its value. I had a lot of trouble with this one. The problem was that I had to wait for the scene to finish loading before starting the enemy spawning. This is now fixed and I added an extra state to the gamemanager. Next up is adding some more enemies and coding some more of their spawning behaviour!