Added pixel shader
Totally not my shader btw. It such a struggle getting this to work. Took 4h before i decided to give up. The problem is that I want to not pixelate the UI. So i tried to create a different camera that will draw the UI. This ended up working in the end but the UI would draw ontop of everything. It is like a never ending loop. If I want only specific game objects to be pixelated, I have to draw it first, and then draw the UI. However, this will result in the UI being drawn over the gameobjects that are pixelated. If I do it the otherway around. I will draw the UI first, and then draw the pixelated gameobjects, but doing it this way will pixelate everything, because the pixelation will pixelate everything that is on the screen! UGH, I'm now trying to just have a world space canvas that will draw the background of the UI, maybe there is another way to do it (there most definitely is) But I'm done for today. Next up is adding some more enemy and weapon things!