Added sell and reroll
This was a surprisingly fast implementation. Had more troubles with the rerolling but it works flawlessly now. Most shop logic is now in place. Maybe I will add a lock mechanic but I'm not sure about it yet. Still need a better way to display information in the shop, have other ui elements. We will get there!